refactor: auth.js를 loginApi.js로 이름 변경 및 import 수정
